Refactor package layout
