+++ README.md
@@ -1,12 +1,17 @@
 # Estat SHP Utils
 
 CSV ファイルを正規化した形で SQLite データベースへ変換するためのツール集です。
-`pandas` と `dbfread` が必要となります。
+`pandas` と `dbfread` が必要となります。プロジェクトを開発環境で利用す
+る場合は次のようにインストールしてください。
+
+```bash
+pip install -e .
+```
 
 ## ディレクトリ構成
 
 
-- `src/` - 汎用ライブラリ
+- `estat_shp_utils/` - 汎用ライブラリ
 - `app/` - コマンドラインスクリプト
 - `dev/` - 開発用ファイル
 - `doc/` - ドキュメント
