+++ pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "estat-shp-utils"
+version = "0.0.0"
+description = "Utilities for converting CSV files to normalized SQLite databases"
+readme = "README.md"
+requires-python = ">=3.8"
+authors = [{name = "Unknown"}]
+dependencies = [
+    "pandas",
+    "dbfread",
+]
+
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["estat_shp_utils*"]
