+++ sample/lookup_city.py
@@ -5,13 +5,9 @@ from __future__ import annotations
 
 import argparse
 from pathlib import Path
-import os
-import sys
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-
-from src.database import Database
-from src.r2ka_api import CityIdSelector
+from estat_shp_utils.database import Database
+from estat_shp_utils.r2ka_api import CityIdSelector
 
 
 def main() -> None:
