+++ sample/lookup_sub_area.py
@@ -5,13 +5,10 @@ from __future__ import annotations
 
 import argparse
 from pathlib import Path
-import os
-import sys
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
-from src.database import Database
-from src.r2ka_api import SubAreaIdSelector
+from estat_shp_utils.database import Database
+from estat_shp_utils.r2ka_api import SubAreaIdSelector
 
 
 def main() -> None:
