+++ tests/test_importer.py
@@ -2,13 +2,10 @@ import sqlite3
 import tempfile
 from pathlib import Path
 import unittest
-import sys
-import os
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
-from src.database import Database
-from src.r2ka_importer import R2KAImporter
+from estat_shp_utils.database import Database
+from estat_shp_utils.r2ka_importer import R2KAImporter
 
 class TestR2KAImporterIntegration(unittest.TestCase):
     def test_db_created_from_dbf(self):
